EIC {flexmix}	R Documentation

Entropic Measure Information Criterion

Description

Compute the Entropic measure Information criterion for model selection.

Usage

## S4 method for signature 'flexmix'
EIC(object, ...)
## S4 method for signature 'stepFlexmix'
EIC(object, ...)

Arguments

object
see Methods section below

...
Some methods for this generic function may take additional, optional arguments. At present none do.

Value

Returns a numeric vector with the corresponding EIC value(s).

Methods

object = "flexmix":

Compute the EIC of a flexmix object.

object = "stepFlexmix":

Compute the EIC of all models contained in the stepFlexmix object.

Author(s)

Bettina Gruen

References

V. Ramaswamy, W. S. DeSarbo, D. J. Reibstein, and W. T. Robinson. An empirical pooling approach for estimating marketing mix elasticities with PIMS data. Marketing Science, 12(1), 103–124, 1993.

Examples

data("NPreg", package = "flexmix")
ex1 <- flexmix(yn ~ x + I(x^2), data = NPreg, k = 2)
EIC(ex1)

[Package flexmix version 2.3-14 Index]