ExNPreg {flexmix}	R Documentation

Artificial Example for Normal, Poisson and Binomial Regression

Description

A simple artificial regression example with 2 latent classes, one independent variable (uniform on [0,10]), and three dependent variables with Gaussian, Poisson and Binomial distribution, respectively.

Usage

ExNPreg(n)
data("NPreg")

Arguments

n
Number of observations per latent class.

Details

The NPreg data frame can be re-created by ExNPreg(100) using set.seed(2602), it has been saved as a data set for simplicity of examples only.

Examples

data("NPreg", package = "flexmix")
plot(yn ~ x, data = NPreg, col = class)
plot(yp ~ x, data = NPreg, col = class)
plot(yb ~ x, data = NPreg, col = class)

[Package flexmix version 2.3-14 Index]