ExNclus {flexmix}	R Documentation

Artificial Example with 4 Gaussians

Description

A simple artificial example for normal clustering with 4 latent classes, all of them having a Gaussian distribution. See the function definition for true means and covariances.

Usage

ExNclus(n)
data("Nclus")

Arguments

n
Number of observations in the two small latent classes.

Details

The Nclus data set can be re-created by ExNclus(100) using set.seed(2602), it has been saved as a data set for simplicity of examples only.

Examples

data("Nclus", package = "flexmix")
require("MASS")
eqscplot(Nclus, col = rep(1:4, c(100, 100, 150, 200)))

[Package flexmix version 2.3-14 Index]