FLXMCdist1 {flexmix}	R Documentation

FlexMix Clustering of Univariate Distributions

Description

These are drivers for flexmix implementing model-based clustering of univariate data using different distributions for the component-specific models.

Usage

FLXMCdist1(formula = . ~ ., dist, ...)

Arguments

formula
A formula which is interpreted relative to the formula specified in the call to flexmix using update.formula. Only the left-hand side (response) of the formula is used. Default is to use the original flexmix model formula.

dist
Character string indicating the component-specific univariate distribution.

...
Arguments for the specific model drivers.

Details

Currently drivers for the following distributions are available:

Lognormal ("lnorm")

inverse Gaussian ("invGauss" using dinvGauss)

gamma ("gamma")

exponential ("exp")

Weibull ("weibull")

Burr ("burr" using dburr)

Inverse Burr ("invburr" using dinvburr)

Value

FLXMCmvnorm returns an object of class FLXMC.

Author(s)

Friedrich Leisch and Bettina Gruen

References

Tatjana Miljkovic and Bettina Gruen. Modeling loss data using mixtures of distributions. Insurance: Mathematics and Economics, 70, 387-396, 2016. doi:10.1016/j.insmatheco.2016.06.019

See Also

flexmix

Examples

if (require("actuar")) {
    set.seed(123)
    y <- c(rexp(100, 10), rexp(100, 1))
    ex <- flexmix(y ~ 1, cluster = rep(1:2, each = 100), model = FLXMCdist1(dist = "exp"))    
    parameters(ex)
}

[Package flexmix version 2.3-14 Index]