FLXMCmvpois {flexmix}	R Documentation

FlexMix Poisson Clustering Driver

Description

This is a model driver for flexmix implementing model-based clustering of Poisson distributed data.

Usage

FLXMCmvpois(formula = . ~ .)

Arguments

formula
A formula which is interpreted relative to the formula specified in the call to flexmix using update.formula. Only the left-hand side (response) of the formula is used. Default is to use the original flexmix model formula.

Details

This can be used to cluster Poisson distributed data where given the component membership the variables are mutually independent.

Value

FLXMCmvpois returns an object of class FLXMC.

Author(s)

Friedrich Leisch and Bettina Gruen

See Also

flexmix

[Package flexmix version 2.3-14 Index]