FLXMRlmmc {flexmix}	R Documentation

FlexMix Interface to Linear Mixed Models with Left-Censoring

Description

This is a driver which allows fitting of mixtures of linear models with random effects and left-censored observations.

Usage

FLXMRlmmc(formula = . ~ ., random, censored, varFix, eps = 10^-6, ...)

Arguments

formula
A formula which is interpreted relative to the formula specified in the call to flexmix using update.formula. Default is to use the original flexmix model formula.

random
A formula for specifying the random effects. If missing no random effects are fitted.

varFix
If random effects are specified a named logical vector of length 2 indicating if the variance of the random effects and the residuals are fixed over the components. Otherwise a logical indicating if the variance of the residuals are fixed over the components.

censored
A formula for specifying the censoring variable.

eps
Observations with an a-posteriori probability smaller or equal to eps are omitted in the M-step.

...
Additional arguments to be passed to lm.wfit.

Value

Returns an object of class FLXMRlmmc, FLXMRlmmcfix, FLXMRlmc or FLXMRlmcfix.

Author(s)

Bettina Gruen

[Package flexmix version 2.3-14 Index]