FLXMRmultinom {flexmix}	R Documentation

FlexMix Interface to Multiomial Logit Models

Description

Model driver for fitting mixtures of multinomial logit models.

Usage

FLXMRmultinom(formula = . ~ ., ...)

Arguments

formula
A formula which is interpreted relative to the formula specified in the call to flexmix using update.formula. Default is to use the original flexmix model formula.

...
Additional arguments to be passed to nnet.default.

Details

The M-step is performed using nnet.default.

Value

Returns an object of class FLXMRmultinom.

Warning

To ensure identifiability repeated measurements are necessary. Sufficient conditions are given in Gruen and Leisch (2008).

Author(s)

Bettina Gruen

References

Bettina Gruen and Friedrich Leisch. Identifiability of finite mixtures of multinomial logit models with varying and fixed effects. Journal of Classification, 25, 225–247. 2008.

See Also

FLXMRcondlogit

[Package flexmix version 2.3-14 Index]