FLXMRziglm {flexmix}	R Documentation

FlexMix Interface to Zero Inflated Generalized Linear Models

Description

This is a driver which allows fitting of zero inflated poisson and binomial models.

Usage

FLXMRziglm(formula = . ~ ., family = c("binomial", "poisson"), ...)

Arguments

formula
A formula which is interpreted relative to the formula specified in the call to flexmix using update.formula. Default is to use the original flexmix model formula.

family
A character string naming a glm family function.

...
passed to FLXMRglm

Value

Returns an object of class FLXMRziglm.

Note

In fact this only approximates zero inflated models by fixing the coefficient of the intercept at -Inf and the other coefficients at zero for the first component.

Author(s)

Friedrich Leisch and Bettina Gruen

Examples

  data("dmft", package = "flexmix")
  Model <- FLXMRziglm(family = "poisson")
  Fitted <- flexmix(End ~ log(Begin + 0.5) + Gender + Ethnic + Treatment, 
                    model = Model, k = 2 , data = dmft, 
                    control = list(minprior = 0.01))
  summary(refit(Fitted))

[Package flexmix version 2.3-14 Index]