FLXP-class {flexmix}	R Documentation

Class "FLXP"

Description

Concomitant model class.

Objects from the Class

Objects can be created by calls of the form new("FLXP", ...), typically inside driver functions like FLXPconstant or FLXPmultinom.

Slots

name:

Character string used in print methods.

formula:

Formula describing the model.

x:

Model matrix.

fit:

Function returning the fitted prior probabilities.

refit:

Function returning the fitted concomitant model.

coef:

Matrix containing the fitted parameters.

df:

Function for determining the number of degrees of freedom used.

Author(s)

Friedrich Leisch and Bettina Gruen

[Package flexmix version 2.3-14 Index]