FLXMCmvbinary {flexmix}	R Documentation

FlexMix Binary Clustering Driver

Description

This is a model driver for flexmix implementing model-based clustering of binary data.

Usage

FLXMCmvbinary(formula = . ~ ., truncated = FALSE)

Arguments

formula
A formula which is interpreted relative to the formula specified in the call to flexmix using update.formula. Only the left-hand side (response) of the formula is used. Default is to use the original flexmix model formula.

truncated
logical, if TRUE the observations for the pattern with only zeros are missing and the truncated likelihood is optimized using an EM-algorithm.

Details

This model driver can be used to cluster binary data. The only parameter is the column-wise mean of the data, which equals the probability of observing a 1.

Value

FLXMCmvbinary returns an object of class FLXMC.

Author(s)

Friedrich Leisch and Bettina Gruen

See Also

flexmix

[Package flexmix version 2.3-14 Index]