FLXP {flexmix}	R Documentation

Creates the Concomitant Variable Model

Description

Creator functions for the concomitant variable model. FLXPconstant specifies constant priors and FLXPmultinom multinomial logit models for the priors.

Usage

FLXPconstant()
FLXPmultinom(formula = ~1)

Arguments

formula
A formula for determining the model matrix of the concomitant variables.

Details

FLXPmultinom uses nnet.default from nnet to fit the multinomial logit model.

Value

Object of class FLXP. FLXPmultinom returns an object of class FLXPmultinom which extends class FLXP directly and is used for method dispatching.

Author(s)

Friedrich Leisch and Bettina Gruen

[Package flexmix version 2.3-14 Index]