FLXcontrol-class {flexmix}	R Documentation

Class "FLXcontrol"

Description

Hyperparameters for the EM algorithm.

Objects from the Class

Objects can be created by calls of the form new("FLXcontrol", ...). In addition, named lists can be coerced to FLXcontrol objects, names are completed if unique (see examples).

Slots

iter.max:

Maximum number of iterations.

minprior:

Minimum prior probability of clusters, components falling below this threshold are removed during the iteration.

tolerance:

The EM algorithm is stopped when the (relative) change of log-likelihood is smaller than tolerance.

verbose:

If a positive integer, then the log-likelihood is reported every verbose iterations. If 0, no output is generated during model fitting.

classify:

Character string, one of "auto", "weighted", "hard" (or "CEM"), "random" or ("SEM").

nrep:

Reports the number of random initializations used in stepFlexmix() to determine the mixture.

Run new("FLXcontrol") to see the default settings of all slots.

Author(s)

Friedrich Leisch and Bettina Gruen

Examples

## have a look at the defaults
new("FLXcontrol")

## corce a list
mycont <- list(iter = 200, tol = 0.001, class = "r")
as(mycont, "FLXcontrol")

[Package flexmix version 2.3-14 Index]