FLXdist {flexmix}	R Documentation

Finite Mixtures of Distributions

Description

Constructs objects of class FLXdist which represent unfitted finite mixture models.

Usage

 FLXdist(formula, k = NULL, model = FLXMRglm(), components,
         concomitant = FLXPconstant())

Arguments

formula
A symbolic description of the model to be fit. The general form is ~x|g where x is the set of predictors and g an optional grouping factor for repeated measurements.

k
Integer specifying the number of cluster or a numeric vector of length equal to the length of components, specifying the prior probabilities of clusters.

model
Object of class FLXM or a list of FLXM objects. Default is the object returned by calling FLXMRglm().

components
A list of length equal to the number of components containing a list of length equal to the number of models which again contains a list of named elements for defining the parameters of the component-specific model.

concomitant
Object of class FLXconcomitant specifying the model for concomitant variables.

Value

Returns an object of class FLXdist.

Author(s)

Bettina Gruen

See Also

FLXdist-class

[Package flexmix version 2.3-14 Index]