FLXdist-class {flexmix}	R Documentation

Class "FLXdist"

Description

Objects of class FLXdist represent unfitted finite mixture models.

Usage

## S4 method for signature 'FLXdist'
parameters(object, component = NULL, model = NULL, which = c("model",
    "concomitant"), simplify = TRUE, drop = TRUE)
## S4 method for signature 'FLXdist'
predict(object, newdata = list(), aggregate = FALSE, ...)

Arguments

object
An object of class "FLXdist".

component
Number of component(s), if NULL all components are returned.

model
Number of model(s), if NULL all models are returned.

which
Specifies if the parameters of the component specific model or the concomitant variable model are returned.

simplify
Logical, if TRUE the returned values are simplified to a vector or matrix if possible.

drop
Logical, if TRUE the function tries to simplify the return object by omitting lists of length one.

newdata
Dataframe containing new data.

aggregate
Logical, if TRUE then the predicted values for each model aggregated over the components are returned.

...
Passed to the method of the model class.

Slots

model

List of FLXM objects.

prior

Numeric vector with prior probabilities of clusters.

components

List describing the components using FLXcomponent objects.

concomitant:

Object of class "FLXP".

formula

Object of class "formula".

call

The function call used to create the object.

k

Number of clusters.

Accessor Functions

The following functions should be used for accessing the corresponding slots:

parameters:

The parameters for each model and component, return value depends on the model.

prior:

Numeric vector of prior class probabilities/component weights

Author(s)

Friedrich Leisch and Bettina Gruen

See Also

FLXdist

[Package flexmix version 2.3-14 Index]