FLXfit {flexmix}	R Documentation

Fitter Function for FlexMix Models

Description

This is the basic computing engine called by flexmix, it should usually not be used directly.

Usage

FLXfit(model, concomitant, control, postunscaled = NULL, groups,
       weights)

Arguments

model
List of FLXM objects.

concomitant
Object of class FLXP.

control
Object of class FLXcontrol.

weights
A numeric vector of weights to be used in the fitting process.

postunscaled
Initial a-posteriori probabilities of the observations at the start of the EM algorithm.

groups
List with components group which is a factor with optional grouping of observations and groupfirst which is a logical vector for the first observation of each group.

Value

Returns an object of class flexmix.

Author(s)

Friedrich Leisch and Bettina Gruen

See Also

flexmix, flexmix-class

[Package flexmix version 2.3-14 Index]