FLXMRglm {flexmix}	R Documentation

FlexMix Interface to Generalized Linear Models

Description

This is the main driver for FlexMix interfacing the glm family of models.

Usage

FLXMRglm(formula = . ~ .,
         family = c("gaussian", "binomial", "poisson", "Gamma"),
         offset = NULL)

Arguments

formula
A formula which is interpreted relative to the formula specified in the call to flexmix using update.formula. Default is to use the original flexmix model formula.

family
A character string naming a glm family function.

offset
This can be used to specify an a priori known component to be included in the linear predictor during fitting.

Details

See flexmix for examples.

Value

Returns an object of class FLXMRglm.

Author(s)

Friedrich Leisch and Bettina Gruen

References

Friedrich Leisch. FlexMix: A general framework for finite mixture models and latent class regression in R. Journal of Statistical Software, 11(8), 2004. doi:10.18637/jss.v011.i08

See Also

flexmix, glm

[Package flexmix version 2.3-14 Index]