FLXMRglmfix {flexmix}	R Documentation

FlexMix Interface to GLMs with Fixed Coefficients

Description

This implements a driver for FlexMix which interfaces the glm family of models and where it is possible to specify fixed (constant) or nested varying coefficients or to ensure that in the Gaussian case the variance estimate is equal for all components.

Usage

FLXMRglmfix(formula = . ~ ., fixed = ~0, varFix = FALSE, nested = NULL, 
       family = c("gaussian", "binomial", "poisson", "Gamma"),
       offset = NULL)

Arguments

formula
A formula which is interpreted relative to the formula specified in the call to flexmix using update.formula. Default is to use the original flexmix model formula.

fixed
A formula which specifies the additional regressors for the fixed (constant) coefficients.

varFix
A logical indicating if the variance estimate for Gaussian components should be constrained to be equal for all components. It can be also a vector specifying the number of components with equal variance.

nested
An object of class FLXnested or a list specifying the nested structure.

family
A character string naming a glm family function.

offset
This can be used to specify an a priori known component to be included in the linear predictor during fitting.

Value

Returns an object of class FLXMRglmfix.

Author(s)

Friedrich Leisch and Bettina Gruen

See Also

FLXMRglm

Examples

data("NPreg", package = "flexmix")
ex <- flexmix(yn ~ x | id2, data = NPreg, k = 2,
              cluster = NPreg$class,
              model = FLXMRglm(yn ~ . + I(x^2)))
ex.fix <- flexmix(yn ~ x | id2, data = NPreg,
                  cluster = posterior(ex),
                  model = FLXMRglmfix(nested = list(k = c(1, 1),
                                      formula = c(~0, ~I(x^2)))))
summary(refit(ex))
## Not run: 
summary(refit(ex.fix))

## End(Not run)

[Package flexmix version 2.3-14 Index]