FLXM-class {flexmix}	R Documentation

Class "FLXM"

Description

FlexMix model specification.

Objects from the Class

Objects can be created by calls of the form new("FLXM", ...), typically inside driver functions like FLXMRglm or FLXMCmvnorm.

Slots

fit:

Function returning an FLXcomponent object.

defineComponent:

Function or expression to determine the FLXcomponent object given the parameters.

weighted:

Logical indicating whether fit can do weighted likelihood maximization.

name:

Character string used in print methods.

formula:

Formula describing the model.

fullformula:

Resulting formula from updating the model formula with the formula specified in the call to flexmix.

x:

Model matrix.

y:

Model response.

terms, xlevels, contrasts:

Additional information for model matrix.

preproc.x:

Function for preprocessing matrix x before the EM algorithm starts, by default the identity function.

preproc.y:

Function for preprocessing matrix y before the EM algorithm starts, by default the identity function.

Author(s)

Friedrich Leisch and Bettina Gruen

[Package flexmix version 2.3-14 Index]