Mehta {flexmix}	R Documentation

Mehta Trial

Description

For a 22-centre trial the number of responses and the total number of patients is reported for the control group and the group receiving a new drug.

Usage

data("Mehta")

Format

A data frame with 44 observations on the following 4 variables.

Response

Number of responses.

Total

Total number of observations.

Drug

A factor indicating treatment with levels New and Control.

Site

A factor indicating the site/centre.

Source

M. Aitkin. Meta-analysis by random effect modelling in generalized linear models. Statistics in Medicine, 18, 2343–2351, 1999.

References

C.R. Mehta, N.R. Patel and P. Senchaudhuri. Importance sampling for estimating exact probabilities in permutational inference. Journal of the American Statistical Association, 83, 999–1005, 1988.

Examples

data("Mehta", package = "flexmix")
mehtaMix <- initFlexmix(cbind(Response, Total-Response) ~ 1|Site, 
                        data = Mehta, nrep = 5, k = 3,
                        model = FLXMRglmfix(family = "binomial",
                          fixed = ~ Drug), 
                        control = list(minprior = 0.04))

[Package flexmix version 2.3-14 Index]