bioChemists {flexmix}	R Documentation

Articles by Graduate Students in Biochemistry Ph.D. Programs

Description

A sample of 915 biochemistry graduate students.

Usage

data("bioChemists")

Format

art

count of articles produced during last 3 years of Ph.D.

fem

factor indicating gender of student, with levels Men and Women

mar

factor indicating marital status of student, with levels Single and Married

kid5

number of children aged 5 or younger

phd

prestige of Ph.D. department

ment

count of articles produced by Ph.D. mentor during last 3 years

Details

This data set is taken from package pscl provided by Simon Jackman.

Source

found in Stata format at http://www.indiana.edu/~jslsoc/stata/spex_data/couart2.dta

References

Long, J. Scott. The origins of sex difference in science. Social Forces, 68, 1297–1315, 1990.

Long, J. Scott. Regression Models for Categorical and Limited Dependent Variables, 1997. Thousand Oaks, California: Sage.

[Package flexmix version 2.3-14 Index]