dmft {flexmix}	R Documentation

Dental Data

Description

Count data from a dental epidemiological study for evaluation of various programs for reducing caries collected among school children from an urban area of Belo Horizonte (Brazil).

Usage

data("dmft")

Format

A data frame with 797 observations on the following 5 variables.

End

Number of decayed, missing or filled teeth at the end of the study.

Begin

Number of decayed, missing or filled teeth at the beginning of the study.

Gender

A factor with levels male and female.

Ethnic

A factor with levels brown, white and black.

Treatment

A factor with levels control, educ, enrich, rinse, hygiene and all.

Details

The aim of the caries prevention study was to compare four methods to prevent dental caries. Interventions were carried out according to the following scheme:

control

Control group

educ

Oral health education

enrich

Enrichment of the school diet with rice bran

rinse

Mouthwash with 0.2% sodium floride (NaF) solution

hygiene

Oral hygiene

all

All four methods together

Source

D. Boehning, E. Dietz, P. Schlattmann, L. Mendonca and U. Kirchner. The zero-inflated Poisson model and the decayed, missing and filled teeth index in dental epidemiology. Journal of the Royal Statistical Society A, 162(2), 195–209, 1999.

Examples

data("dmft", package = "flexmix")
dmft_flx <- initFlexmix(End ~ 1, data = dmft, k = 2,
                        model = FLXMRglmfix(family = "poisson", 
                        fixed = ~ Gender + Ethnic + Treatment))

[Package flexmix version 2.3-14 Index]