fabricfault {flexmix}	R Documentation

Fabric Faults

Description

Number of faults in rolls of a textile fabric.

Usage

data("fabricfault")

Format

A data frame with 32 observations on the following 2 variables.

Length

Length of role (m).

Faults

Number of faults.

Source

G. McLachlan and D. Peel. Finite Mixture Models, 2000, John Wiley and Sons Inc. http://www.maths.uq.edu.au/~gjm/DATA/mmdata.html

References

A. F. Bissell. A Negative Binomial Model with Varying Element Sizes Biometrika, 59, 435–441, 1972.

M. Aitkin. A general maximum likelihood analysis of overdispersion in generalized linear models. Statistics and Computing, 6, 251–262, 1996.

Examples

data("fabricfault", package = "flexmix")
fabricMix <- initFlexmix(Faults ~ 1, data = fabricfault, k = 2,
                         model = FLXMRglmfix(family = "poisson",
                           fixed = ~ log(Length)), 
                         nrep = 5)

[Package flexmix version 2.3-14 Index]