flexmix-class {flexmix}	R Documentation

Class "flexmix"

Description

A fitted flexmix model.

Slots

model:

List of FLXM objects.

prior:

Numeric vector with prior probabilities of clusters.

posterior:

Named list with elements scaled and unscaled, both matrices with one row per observation and one column per cluster.

iter:

Number of EM iterations.

k:

Number of clusters after EM.

k0:

Number of clusters at start of EM.

cluster:

Cluster assignments of observations.

size:

Cluster sizes.

logLik:

Log-likelihood at EM convergence.

df:

Total number of parameters of the model.

components:

List describing the fitted components using FLXcomponent objects.

formula:

Object of class "formula".

control:

Object of class "FLXcontrol".

call:

The function call used to create the object.

group:

Object of class "factor".

converged:

Logical, TRUE if EM algorithm converged.

concomitant:

Object of class "FLXP"..

weights:

Optional weights of the observations.

Extends

Class FLXdist, directly.

Accessor Functions

The following functions should be used for accessing the corresponding slots:

cluster:

Cluster assignments of observations.

posterior:

A matrix of posterior probabilities for each observation.

Author(s)

Friedrich Leisch and Bettina Gruen

[Package flexmix version 2.3-14 Index]