patent {flexmix}	R Documentation

Patents and R&D Spending

Description

Number of patents, R&D spending and sales in millions of dollar for 70 pharmaceutical and biomedical companies in 1976.

Usage

data("patent")

Format

A data frame with 70 observations on the following 4 variables.

Company

Name of company.

Patents

Number of patents.

RDS

R&D spending per sales.

lgRD

Logarithmized R&D spendings (in millions of dollars).

Details

The data is taken from the National Bureau of Economic Research R\&D Masterfile.

Source

P. Wang, I.M. Cockburn and M.L. Puterman. Analysis of Patent Data – A Mixed-Poisson-Regression-Model Approach. Journal of Business & Economic Statistics, 16(1), 27–41, 1998.

References

B.H. Hall, C. Cummins, E. Laderman and J. Mundy. The R&D Master File Documentation. Technical Working Paper 72, National Bureau of Economic Research, 1988. Cambridge, MA.

Examples

data("patent", package = "flexmix")
patentMix <- initFlexmix(Patents ~ lgRD, k = 3,
                         model = FLXMRglm(family = "poisson"),
                         concomitant = FLXPmultinom(~RDS),
                         nrep = 5, data = patent)
plot(Patents ~ lgRD, data = patent,
     pch = as.character(clusters(patentMix)))
ordering <- order(patent$lgRD)
apply(fitted(patentMix), 2, function(y)
      lines(sort(patent$lgRD), y[ordering]))

[Package flexmix version 2.3-14 Index]