plotEll {flexmix}	R Documentation

Plot Confidence Ellipses for FLXMCmvnorm Results

Description

Plot 50% and 95% confidence ellipses for mixtures of Gaussians fitted using FLXMCmvnorm.

Usage

plotEll(object, data, which = 1:2, model = 1, project = NULL, points = TRUE,
        eqscale = TRUE, col = NULL, number = TRUE, cex = 1.5, numcol = "black", 
        pch = NULL, ...)

Arguments

object
An object of class flexmix with a fitted FLXMCmvnorm model.

data
The response variable in a data frame or as a matrix.

which
Index numbers of dimensions of (projected) input space to plot.

model
The model (for a multivariate response) that shall be plotted.

project
Projection object, currently only the result of prcomp is supported.

points
Logical, shall data points be plotted?

eqscale
Logical, plot using eqscplot?

number
Logical, plot number labels at cluster centers?

cex, numcol
Size and color of number labels.

pch, col, ...
Graphical parameters.

Author(s)

Friedrich Leisch and Bettina Gruen

See Also

FLXMCmvnorm

[Package flexmix version 2.3-14 Index]