posterior {flexmix}	R Documentation

Determine Cluster Membership and Posterior Probabilities

Description

Determine posterior probabilities or cluster memberships for a fitted flexmix or unfitted FLXdist model.

Usage

## S4 method for signature 'flexmix,missing'
posterior(object, newdata, unscaled = FALSE, ...)
## S4 method for signature 'FLXdist,listOrdata.frame'
posterior(object, newdata, unscaled = FALSE, ...)
## S4 method for signature 'flexmix,missing'
clusters(object, newdata, ...)
## S4 method for signature 'FLXdist,ANY'
clusters(object, newdata, ...)

Arguments

object
An object of class "flexmix" or "FLXdist".

newdata
Data frame or list containing new data. If missing the posteriors of the original observations are returned.

unscaled
Logical, if TRUE the component-specific likelihoods are returned.

...
Currently not used.

Author(s)

Friedrich Leisch and Bettina Gruen

[Package flexmix version 2.3-14 Index]