salmonellaTA98 {flexmix}	R Documentation

Salmonella Reverse Mutagenicity Assay

Description

Data on Ames Salmonella reverse mutagenicity assay.

Usage

data("salmonellaTA98")

Format

This data frame contains the following columns:

x

Dose levels of quinoline.

y

Numbers of revertant colonies of TA98 Salmonella observed on each of three replicate plates tested at each of six dose levels of quinoline diameter.

Details

This data set is taken from package dispmod provided by Luca Scrucca.

Source

Margolin, B.J., Kaplan, N. and Zeiger, E. Statistical analysis of the Ames Salmonella/microsome test, Proc. Natl. Acad. Sci. USA, 76, 3779–3783, 1981.

References

Breslow, N.E. Extra-Poisson variation in log-linear models, Applied Statistics, 33, 38–44, 1984.

Wang, P., Puterman, M.L., Cockburn, I.M., and Le, N.D. Mixed Poisson regression models with covariate dependent rates, Biometrics, 52, 381–400, 1996.

Examples

data("salmonellaTA98", package = "flexmix")
salmonMix <- initFlexmix(y ~ 1,
                         data = salmonellaTA98, 
                         model = FLXMRglmfix(family = "poisson", 
                           fixed = ~ x + log(x + 10)),                        
                         k = 2, nrep = 5)
salmonMix.pr <- predict(salmonMix, newdata = salmonellaTA98)
plot(y ~ x, data = salmonellaTA98, 
     pch = as.character(clusters(salmonMix)), 
     ylim = range(c(salmonellaTA98$y, unlist(salmonMix.pr))))
for (i in 1:2) lines(salmonellaTA98$x, salmonMix.pr[[i]], lty = i)

[Package flexmix version 2.3-14 Index]