tribolium {flexmix}	R Documentation

Tribolium Beetles

Description

The data investigates whether the adult Tribolium species Castaneum has developed an evolutionary advantage to recognize and avoid eggs of their own species while foraging.

Usage

data("tribolium")

Format

A data frame with 27 observations on the following 4 variables.

Remaining

A numeric vector.

Total

A numeric vector.

Replicate

A factor with levels 1, 2, 3.

Species

A factor with levels Castaneum Confusum Madens.

Details

Beetles of the genus Tribolium are cannibalistic in the sense that adults eat the eggs of their own species as well as those of closely related species. The experiment isolated a number of adult beetles of the same species and presented them with a vial of 150 eggs (50 of each type), the eggs being thoroughly mixed to ensure uniformity throughout the vial.

The data gives the consumption data for adult Castaneum species. It reports the number of Castaneum, Confusum and Madens eggs, respectively, that remain uneaten after two day exposure to the adult beetles. Replicates 1, 2, and 3 correspond to different occasions on which the experiment was conducted.

Source

P. Wang and M.L. Puterman. Mixed Logistic Regression Models. Journal of Agricultural, Biological, and Environmental Statistics, 3 (2), 175–200, 1998.

Examples

data("tribolium", package = "flexmix")
tribMix <- initFlexmix(cbind(Remaining, Total - Remaining) ~ Species, 
                   k = 2, nrep = 5, data = tribolium,
                   model = FLXMRglm(family = "binomial"))

[Package flexmix version 2.3-14 Index]