trypanosome {flexmix}	R Documentation

Trypanosome

Description

Trypanosome data from a dosage-response analysis to assess the proportion of organisms belonging to different populations. It is assumed that organisms belonging to different populations are indistinguishable other than in terms of their reaction to the stimulus.

Usage

data("trypanosome")

Format

A data frame with 426 observations on the following 2 variables.

Dead

A logical vector.

Dose

A numeric vector.

Details

The experimental technique involved inspection under the microscope of a representative aliquot of a suspension, all organisms appearing within two fields of view being classified either alive or dead. Hence the total numbers of organisms present at each dose and the number showing the quantal response were both random variables.

Source

R. Ashford and P.J. Walker. Quantal Response Analysis for a Mixture of Populations. Biometrics, 28, 981–988, 1972.

References

D.A. Follmann and D. Lambert. Generalizing Logistic Regression by Nonparametric Mixing. Journal of the American Statistical Association, 84(405), 195–300, 1989.

Examples

data("trypanosome", package = "flexmix")
trypMix <- initFlexmix(cbind(Dead, 1-Dead) ~ 1, k = 2, 
	               nrep = 5, data = trypanosome,
	               model = FLXMRglmfix(family = "binomial",
                           fixed = ~log(Dose)))

[Package flexmix version 2.3-14 Index]