whiskey {flexmix}	R Documentation

Survey Data on Brands of Scotch whiskey Consumed

Description

The data set is from Simmons Study of Media and Markets and contains the incidence matrix for scotch brands used in last year for those households who report consuming scotch.

Usage

data("whiskey")

Format

A data frame whiskey with 484 observations on the following 2 variables.

Freq

a numeric vector

Incidence

a matrix with 21 columns

Additional information on the brands is contained in the data frame whiskey_brands which is simultaneously loaded. This data frame contains 21 observations on the following 3 variables.

Brand

a character vector

Type

a factor with levels Blend Single Malt

Bottled

a factor with levels Domestic Foreign

Details

The dataset is taken from the bayesm package.

Source

Peter Rossi and Rob McCulloch. bayesm: Bayesian Inference for Marketing/Micro-econometrics. R package version 2.0-8, 2006. http://gsbwww.uchicago.edu/fac/peter.rossi/research/bsm.html

References

Edwards, Y. and G. Allenby. Multivariate Analysis of Multiple Response Data, Journal of Marketing Research, 40, 321–334, 2003.

[Package flexmix version 2.3-14 Index]